Separate datasource and sentinel APIs
