+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/DataSourceService.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.api;
+package org.bithon.server.web.service.datasource.api;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
