+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/impl/DataSourceApi.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.api;
+package org.bithon.server.web.service.datasource.api.impl;
 
 import lombok.Data;
 import org.bithon.component.commons.utils.CollectionUtils;
@@ -34,6 +34,17 @@ import org.bithon.server.storage.metrics.Interval;
 import org.bithon.server.storage.metrics.ListQuery;
 import org.bithon.server.storage.metrics.MetricStorageConfig;
 import org.bithon.server.storage.metrics.TimeseriesQueryV2;
+import org.bithon.server.web.service.datasource.api.DataSourceService;
+import org.bithon.server.web.service.datasource.api.DisplayableText;
+import org.bithon.server.web.service.datasource.api.GeneralQueryRequest;
+import org.bithon.server.web.service.datasource.api.GetDimensionRequest;
+import org.bithon.server.web.service.datasource.api.GetDimensionRequestV2;
+import org.bithon.server.web.service.datasource.api.GroupByQueryRequest;
+import org.bithon.server.web.service.datasource.api.IDataSourceApi;
+import org.bithon.server.web.service.datasource.api.ListQueryRequest;
+import org.bithon.server.web.service.datasource.api.ListQueryResponse;
+import org.bithon.server.web.service.datasource.api.TimeSeriesQueryRequest;
+import org.bithon.server.web.service.datasource.api.UpdateTTLRequest;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
