+++ server/web-service/src/main/java/org/bithon/server/web/service/sentinel/CreateFlowRuleRequest.java
@@ -14,18 +14,22 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.api;
+package org.bithon.server.web.service.sentinel;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
+import lombok.Data;
+import org.bithon.agent.sentinel.flow.FlowRuleDto;
+
+import javax.validation.constraints.NotNull;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2021/3/6 1:35 下午
+ * @date 2021/7/6 7:19 下午
  */
-@Getter
-@AllArgsConstructor
-public class DisplayableText {
-    private final String value;
-    private final String text;
+@Data
+public class CreateFlowRuleRequest extends FlowRuleDto {
+    /**
+     * application name with profile
+     */
+    @NotNull
+    private String appName;
 }
